refactor: overlay 파일 분리
