+++ client/front/lib/screens/main/errand_list/check_errand/fix_errand/fix_errand.dart
@@ -7,6 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:flutter_svg/svg.dart';
+import 'package:front/widgets/bar/navigation_bar.dart';
 import '../check_errand.dart';
 import '../../../write_errand/write_errand.dart';
 import '../../errand_list.dart';
@@ -16,115 +17,7 @@ import 'fix_errand_widget/fixminimap.dart';
 import 'package:http/http.dart' as http;
 final storage = FlutterSecureStorage();
 
-void _insertOverlay(BuildContext context) {
-  if (overlayEntry != null) return;
-  overlayEntry = OverlayEntry(
-    builder: (context) => Positioned(
-      bottom: 0.h,
-      left: 0.w,
-      right: 0.w,
-      child: Container(
-        width: 364.w,
-        height: 64.h,
-        decoration: BoxDecoration(
-          color: Color(0xffFFFFFF),
-          border: Border.all(
-            color: Color(0xffCFCFCF),
-            width: 0.5.w,
-          ),
-          boxShadow: [
-            BoxShadow(
-              color: Color.fromRGBO(185, 185, 185, 0.25),
-              offset: Offset(5.w, -1.h),
-              blurRadius: 5,
-              spreadRadius: 1,
-            ),
-          ],
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Container(
-              width: 22.w,
-              height: 22.h,
-              margin: EdgeInsets.only(left: 44.w, top: 20.0.h, bottom: 17.32.h),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {},
-                icon: SvgPicture.asset(
-                  'assets/images/home_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-            Container(
-              width: 19.31.w,
-              height: 23.81.h,
-              margin: EdgeInsets.only(top: 20.0.h, bottom: 17.32.h),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {},
-                icon: SvgPicture.asset(
-                  'assets/images/profile_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-            Container(
-              width: 22.0.w,
-              height: 22.h,
-              margin: EdgeInsets.only(top: 20.0.h, bottom: 17.32.h),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {
-                  Navigator.of(context).push(
-                    MaterialPageRoute(
-                      builder: (BuildContext context) => WriteErrand(),
-                    ),
-                  );
-                },
-                icon: SvgPicture.asset(
-                  'assets/images/write_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-            Container(
-              width: 21.95.w,
-              height: 24.21.h,
-              margin: EdgeInsets.only(top: 20.0.h, bottom: 17.32.h, right: 43.92.w),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {},
-                icon: SvgPicture.asset(
-                  'assets/images/history_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    ),
-  );
-  Overlay.of(context).insert(overlayEntry!);
-}
+
 class FixErrand extends StatefulWidget {
   final Map<String, dynamic> errands;
 
@@ -165,7 +58,7 @@ class _FixErrandState extends State<FixErrand> {
             "Content-Type": "application/json"
           });
       if (response.statusCode == 200) {
-        _insertOverlay(context);
+        insertOverlay(context);
         Navigator.of(context).push(
           MaterialPageRoute(
               builder: (context) => MainErrandCheck(errandNo: errandNo,)
@@ -341,7 +234,7 @@ class _FixErrandState extends State<FixErrand> {
           canPop: true,
           onPopInvoked: (bool didPop) async {
         if (didPop) {
-          _insertOverlay(context);
+          insertOverlay(context);
           return;
         }
         Navigator.of(context).pop();
