+++ client/front/lib/screens/main/write_errand/write_errand.dart
@@ -9,129 +9,15 @@ import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_naver_map/flutter_naver_map.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:front/widgets/bar/navigation_bar.dart';
 import 'package:geolocator/geolocator.dart';
 import 'find_destination.dart';
 import 'package:http/http.dart' as http;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import '../errand_list/check_errand/check_errand.dart';
 import '../errand_list/errand_list.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
 final storage = FlutterSecureStorage();
-void _insertOverlay(BuildContext context) {
 
-  if (overlayEntry != null) return;
-  overlayEntry = OverlayEntry(
-    builder: (context) => Positioned(
-      bottom: 0.h,
-      left: 0.w,
-      right: 0.w,
-      child: Container(
-        width: 360.w,
-        height: 64.h,
-        decoration: BoxDecoration(
-          color: Color(0xffFFFFFF),
-          border: Border.all(
-            color: Color(0xffCFCFCF),
-            width: 0.5.w,
-          ),
-          boxShadow: [
-            BoxShadow(
-              color: Color.fromRGBO(185, 185, 185, 0.25),
-              offset: Offset(5, -1),
-              blurRadius: 5,
-              spreadRadius: 1,
-            ),
-          ],
-        ),
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-          children: [
-            Container(
-              width: 22.w,
-              height: 22.h,
-              margin: EdgeInsets.only(left: 44.w,  top: 20.h, bottom: 17.32.h, ),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(builder: (context) => Home()));
-                },
-                icon: SvgPicture.asset(
-                  'assets/images/home_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-            Container(
-              width: 19.31.w,
-              height: 23.81.h,
-              margin: EdgeInsets.only(top: 20.h, bottom: 17.32.h,),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {},
-                icon: SvgPicture.asset(
-                  'assets/images/profile_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-            Container(
-              width: 22.0.w,
-              height: 22.h,
-              margin: EdgeInsets.only(top: 20.h, bottom: 17.32.h,),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {
-                  Navigator.of(context).push(
-                    MaterialPageRoute(
-                      builder: (BuildContext context) => WriteErrand(),
-                    ),
-                  );
-                },
-                icon: SvgPicture.asset(
-                  'assets/images/write_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-            Container(
-              width: 21.95.w,
-              height: 24.21.h,
-              margin: EdgeInsets.only(top: 20.h, bottom: 17.32.h,
-                right: 43.92.w, ),
-              child: IconButton(
-                style: IconButton.styleFrom(
-                  minimumSize: Size.zero,
-                  padding: EdgeInsets.zero,
-                  tapTargetSize: MaterialTapTargetSize.shrinkWrap,
-                ),
-                onPressed: () {},
-                icon: SvgPicture.asset(
-                  'assets/images/history_icon.svg',
-                  color: Color(0xffADADAD),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    ),
-  );
-  Overlay.of(context).insert(overlayEntry!);
-}
 //현재 화면에서 뒤로가기
 class WriteErrand extends StatefulWidget {
   WriteErrand({super.key});
@@ -273,7 +159,7 @@ class _WriteErrandState extends State<WriteErrand> {
           });
       if (response.statusCode == 200) {
         int errandNo = jsonDecode(response.body)['errandNo'];
-        _insertOverlay(context);
+        insertOverlay(context);
         Navigator.of(context).push(
             MaterialPageRoute(
                 builder: (context) => MainErrandCheck(errandNo: errandNo,)
@@ -405,7 +291,7 @@ class _WriteErrandState extends State<WriteErrand> {
         canPop: true,
         onPopInvoked: (bool didPop) async {
       if (didPop) {
-        _insertOverlay(context);
+        insertOverlay(context);
         return;
       }
       Navigator.of(context).pop();
